+++ app/Capture.hs
@@ -1,7 +1,10 @@
+{-# OPTIONS_GHC -Wno-name-shadowing #-}
+
 module Capture where
 
 import Content
 import Data.Fixed (Micro)
+import Data.Foldable (for_)
 import Grim qualified
 import Region
 import WfRecorder qualified
@@ -16,11 +19,31 @@ data CaptureArgs = CaptureArgs
     { captureAction :: CaptureAction
     , delay :: Maybe Micro
     , duration :: Maybe Micro
+    , cursor :: Bool
+    , audio :: Bool
+    , audioDevice :: Maybe String
+    , format :: Maybe ContentType
+    , quality :: Maybe Int
+    , scale :: Maybe Double
+    , framerate :: Maybe Int
     }
 
 capture :: CaptureArgs -> Region -> IO Content
 capture CaptureArgs{..} region = do
-    mapM_ (countdown "Starting in: ") delay
+    for_ delay $ countdown "Starting in: "
     case captureAction of
-        Screenshot -> png <$> Grim.screenshotRegion region
-        Video -> mp4 <$> WfRecorder.recordRegion duration region
+        Screenshot -> do
+            contentType <- validateFormat [PNG, JPEG, PPM] PNG
+            content <- Grim.screenshotRegion contentType cursor quality scale region
+            pure Content{..}
+        Video -> do
+            contentType <- validateFormat [MP4, WEBM] MP4
+            content <-
+                WfRecorder.recordRegion contentType audio audioDevice framerate duration region
+            pure Content{..}
+  where
+    validateFormat :: [ContentType] -> ContentType -> IO ContentType
+    validateFormat allowed def = case format of
+        Nothing -> pure def
+        Just fmt | fmt `elem` allowed -> pure fmt
+        Just _ -> fatalError "Invalid --format for this capture mode"