+++ app/Main.hs
@@ -5,7 +5,7 @@ import Capture
 import Control.Concurrent.Extra
 import Control.Exception.Extra
 import Control.Monad (join)
-import Persistence (createNamedRegionTable)
+import Persistence (createTables)
 import Selection
 import Sink
 import System.Console.ANSI
@@ -20,7 +20,7 @@ import Prelude
 
 main :: IO ()
 main = do
-    createNamedRegionTable
+    createTables
     args <- runParser
     hHideCursor stderr
     hSetBuffering stdout NoBuffering